rename file/dir Signed-off-by: Naman Anand <naman.anand@nimbleedgehq.ai>
