--- sdks/android/app/src/main/kotlin/ai/nimbleedge/android/sampleapp/MainApplication.kt
@@ -1,11 +1,0 @@
-/*
- * SPDX-FileCopyrightText: (C) 2025 DeliteAI Authors
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package ai.nimbleedge.android.sampleapp
-
-import android.app.Application
-
-open class MainApplication : Application()
