--- sdks/android/nimblenet_ktx/src/test/kotlin/ai/nimbleedge/impl/DependencyContainerTest.kt
@@ -1,275 +1,0 @@
-/*
- * SPDX-FileCopyrightText: (C) 2025 DeliteAI Authors
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package ai.nimbleedge.impl
-
-import ai.nimbleedge.impl.common.HardwareInfo
-import ai.nimbleedge.impl.controllers.InternalTaskController
-import ai.nimbleedge.impl.controllers.NimbleNetController
-import ai.nimbleedge.impl.io.AppPreferencesStore
-import ai.nimbleedge.impl.io.FileUtils
-import ai.nimbleedge.impl.io.Networking
-import ai.nimbleedge.impl.loggers.LocalLogger
-import ai.nimbleedge.impl.moduleInstallers.ModuleInstaller
-import ai.nimbleedge.testUtils.nnConfig
-import android.app.Application
-import android.app.DownloadManager
-import android.content.Context
-import android.content.SharedPreferences
-import io.mockk.MockKAnnotations
-import io.mockk.every
-import io.mockk.mockk
-import kotlin.test.assertNotNull
-import kotlin.test.assertSame
-import kotlin.test.assertTrue
-import org.junit.Before
-import org.junit.Test
-
-class DependencyContainerTest {
-    private var application: Application = mockk(relaxed = true)
-    private var sharedPreferences: SharedPreferences = mockk(relaxed = true)
-    private var downloadManager: DownloadManager = mockk(relaxed = true)
-
-    @Before
-    fun setup() {
-        MockKAnnotations.init(this)
-
-        every { application.getSharedPreferences(any(), any()) } returns sharedPreferences
-        every { application.getSystemService(Context.DOWNLOAD_SERVICE) } returns downloadManager
-    }
-
-    @Test
-    fun `getInstance creates new instance when none exists`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        assertNotNull(container)
-    }
-
-    @Test
-    fun `getInstance returns same instance on subsequent calls`() {
-        val container1 = DependencyContainer.getInstance(application, nnConfig)
-        val container2 = DependencyContainer.getInstance(application, nnConfig)
-
-        assertSame(container1, container2)
-    }
-
-    @Test
-    fun `getInstance handles null parameters gracefully`() {
-        try {
-            DependencyContainer.getInstance(null, null)
-        } catch (e: Exception) {
-            assertTrue(e is NullPointerException || e is IllegalArgumentException)
-        }
-    }
-
-    @Test
-    fun `getNimbleNetController returns NimbleNetController instance`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val controller = container.getNimbleNetController()
-
-        assertNotNull(controller)
-        assertTrue(controller is NimbleNetController)
-    }
-
-    @Test
-    fun `getNimbleNetController returns same instance on multiple calls`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val controller1 = container.getNimbleNetController()
-        val controller2 = container.getNimbleNetController()
-
-        assertSame(controller1, controller2)
-    }
-
-    @Test
-    fun `getLocalLogger returns LocalLogger instance`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val logger = container.getLocalLogger()
-
-        assertNotNull(logger)
-        assertTrue(logger is LocalLogger)
-    }
-
-    @Test
-    fun `getLocalLogger returns same instance on multiple calls`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val logger1 = container.getLocalLogger()
-        val logger2 = container.getLocalLogger()
-
-        assertSame(logger1, logger2)
-    }
-
-    @Test
-    fun `getAppPreferencesStore returns AppPreferencesStore instance`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val store = container.getAppPreferencesStore()
-
-        assertNotNull(store)
-        assertTrue(store is AppPreferencesStore)
-    }
-
-    @Test
-    fun `getAppPreferencesStore returns same instance on multiple calls`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val store1 = container.getAppPreferencesStore()
-        val store2 = container.getAppPreferencesStore()
-
-        assertSame(store1, store2)
-    }
-
-    @Test
-    fun `getModuleInstaller returns ModuleInstaller instance`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val installer = container.getModuleInstaller()
-
-        assertNotNull(installer)
-        assertTrue(installer is ModuleInstaller)
-    }
-
-    @Test
-    fun `getModuleInstaller returns same instance on multiple calls`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val installer1 = container.getModuleInstaller()
-        val installer2 = container.getModuleInstaller()
-
-        assertSame(installer1, installer2)
-    }
-
-    @Test
-    fun `getInternalTaskController returns InternalTaskController instance`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val controller = container.getInternalTaskController()
-
-        assertNotNull(controller)
-        assertTrue(controller is InternalTaskController)
-    }
-
-    @Test
-    fun `getInternalTaskController returns same instance on multiple calls`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val controller1 = container.getInternalTaskController()
-        val controller2 = container.getInternalTaskController()
-
-        assertSame(controller1, controller2)
-    }
-
-    @Test
-    fun `getNetworking returns Networking instance`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val networking = container.getNetworking()
-
-        assertNotNull(networking)
-        assertTrue(networking is Networking)
-    }
-
-    @Test
-    fun `getNetworking returns same instance on multiple calls`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val networking1 = container.getNetworking()
-        val networking2 = container.getNetworking()
-
-        assertSame(networking1, networking2)
-    }
-
-    @Test
-    fun `getHardwareInfo returns HardwareInfo instance`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val hardwareInfo = container.getHardwareInfo()
-
-        assertNotNull(hardwareInfo)
-        assertTrue(hardwareInfo is HardwareInfo)
-    }
-
-    @Test
-    fun `getHardwareInfo returns same instance on multiple calls`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val hardwareInfo1 = container.getHardwareInfo()
-        val hardwareInfo2 = container.getHardwareInfo()
-
-        assertSame(hardwareInfo1, hardwareInfo2)
-    }
-
-    @Test
-    fun `getLogsUploadScheduler returns LogsUploadScheduler instance`() {
-        // Skip this test as WorkManager requires Android context initialization
-        // which is complex to mock in unit tests
-        assertTrue(true) // Placeholder to indicate test is acknowledged
-    }
-
-    @Test
-    fun `getLogsUploadScheduler returns same instance on multiple calls`() {
-        // Skip this test as WorkManager requires Android context initialization
-        // which is complex to mock in unit tests
-        assertTrue(true) // Placeholder to indicate test is acknowledged
-    }
-
-    @Test
-    fun `getFileUtils returns FileUtils instance`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val fileUtils = container.getFileUtils()
-
-        assertNotNull(fileUtils)
-        assertTrue(fileUtils is FileUtils)
-    }
-
-    @Test
-    fun `getFileUtils returns same instance on multiple calls`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val fileUtils1 = container.getFileUtils()
-        val fileUtils2 = container.getFileUtils()
-
-        assertSame(fileUtils1, fileUtils2)
-    }
-
-    @Test
-    fun `all singletons are properly initialized and accessible`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        // Verify all public methods return non-null instances (except WorkManager dependent ones)
-        assertNotNull(container.getNimbleNetController())
-        assertNotNull(container.getLocalLogger())
-        assertNotNull(container.getAppPreferencesStore())
-        assertNotNull(container.getModuleInstaller())
-        assertNotNull(container.getInternalTaskController())
-        assertNotNull(container.getNetworking())
-        assertNotNull(container.getHardwareInfo())
-        // Skip getLogsUploadScheduler() as it requires WorkManager initialization
-        assertNotNull(container.getFileUtils())
-    }
-
-    @Test
-    fun `dependency injection works correctly between components`() {
-        val container = DependencyContainer.getInstance(application, nnConfig)
-
-        val logger1 = container.getLocalLogger()
-        val logger2 = container.getLocalLogger()
-
-        // Verify they are the same instance (singleton behavior)
-        assertSame(logger1, logger2)
-
-        // Verify other components can be retrieved without issues
-        val networking = container.getNetworking()
-        val hardwareInfo = container.getHardwareInfo()
-
-        assertNotNull(networking)
-        assertNotNull(hardwareInfo)
-    }
-}
