--- sdks/android/nimblenet_ktx/src/test/kotlin/ai/nimbleedge/testUtils/Reflection.kt
@@ -1,19 +1,0 @@
-/*
- * SPDX-FileCopyrightText: (C) 2025 DeliteAI Authors
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package ai.nimbleedge.testUtils
-
-import kotlin.reflect.full.declaredFunctions
-import kotlin.reflect.jvm.isAccessible
-
-object Reflection {
-    inline fun <reified T : Any, reified R> callPrivateMethod(instance: T, methodName: String): R {
-        val kfun =
-            T::class.declaredFunctions.first { it.name == methodName }.apply { isAccessible = true }
-        @Suppress("UNCHECKED_CAST")
-        return kfun.call(instance) as R
-    }
-}
