+++ sdks/android/nimblenet_ktx/src/test/kotlin/dev/deliteai/testUtils/Constants.kt
@@ -1,0 +1,22 @@
+/*
+ * SPDX-FileCopyrightText: (C) 2025 DeliteAI Authors
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package dev.deliteai.testUtils
+
+import dev.deliteai.datamodels.NimbleNetConfig
+import dev.deliteai.impl.common.NIMBLENET_VARIANTS
+
+val nnConfig =
+    NimbleNetConfig(
+        clientId = "test",
+        host = "test",
+        deviceId = "test",
+        clientSecret = "test",
+        debug = true,
+        initTimeOutInMs = 1000000000,
+        compatibilityTag = "android-output-verification",
+        libraryVariant = NIMBLENET_VARIANTS.STATIC,
+    )
